+++ elements/rh-footer/rh-footer-social-link.ts
@@ -1,5 +1,6 @@
 import { LitElement, html } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
+import { customElement } from 'lit/decorators/custom-element.js';
+import { property } from 'lit/decorators/property.js';
 import { Logger } from '@patternfly/pfe-core/controllers/logger.js';
 
 import style from './rh-footer-social-link.css';
@@ -33,7 +34,7 @@ export class RhFooterSocialLink extends LitElement {
       newDiv.querySelectorAll('[_rendered]').forEach(i => i.remove());
       // NB: icons are restricted to fab set, so as not to require a minor release
       // rh-icon is slated to deal with this problem in-house
-      newDiv.innerHTML = `<pfe-icon icon="${this.icon}" set="fab" loading="eager">${newDiv.innerHTML}</pfe-icon>`;
+      newDiv.innerHTML = `<pf-icon icon="${this.icon}" set="fab" loading="eager">${newDiv.innerHTML}</pf-icon>`;
       // add a11y settings
       /** @todo add logging that warns the user there is an empty label */
       newDiv.setAttribute('aria-label', newDiv.textContent || '');