+++ elements/rh-footer/rh-footer.ts
@@ -1,6 +1,17 @@
+import './rh-footer-social-link.js';
+import './rh-footer-links.js';
+import './rh-footer-block.js';
+import './rh-global-footer.js';
+
+import '@patternfly/elements/pf-icon/pf-icon.js';
+import '@rhds/elements/rh-accordion/rh-accordion.js';
+
 import { RhFooter } from './RhFooter.js';
+import { RhGlobalFooter } from './rh-global-footer.js';
 
 customElements.define('rh-footer', RhFooter);
+
+export { RhFooter, RhGlobalFooter };
 
 declare global {
   interface HTMLElementTagNameMap {