+++ elements/rh-stat/rh-stat.ts
@@ -1,15 +1,16 @@
-import type { ColorTheme } from '../../lib/context/color.js';
-
 import { LitElement, html } from 'lit';
-import { customElement, property } from 'lit/decorators.js';
+import { customElement } from 'lit/decorators/custom-element.js';
+import { property } from 'lit/decorators/property.js';
 import { classMap } from 'lit/directives/class-map.js';
 
-import { colorContextConsumer } from '@patternfly/pfe-core/decorators.js';
+import { colorContextConsumer, type ColorTheme } from '../../lib/context/color/consumer.js';
+
 import { SlotController } from '@patternfly/pfe-core/controllers/slot-controller.js';
 import { Logger } from '@patternfly/pfe-core/controllers/logger.js';
 import { ScreenSizeController } from '../../lib/ScreenSizeController.js';
 
 import styles from './rh-stat.css';
+import { ifDefined } from 'lit/directives/if-defined.js';
 
 /**
  * A statistic showcases a data point or quick fact in a way that visually stands out.
@@ -29,8 +30,7 @@ export class RhStat extends LitElement {
 
   static readonly styles = [styles];
 
-  @colorContextConsumer()
-  @property({ reflect: true }) on: ColorTheme = 'light';
+  @colorContextConsumer() private on?: ColorTheme;
 
   @property({ reflect: true, type: String }) icon?: string;
 
@@ -61,11 +61,14 @@ export class RhStat extends LitElement {
     const hasStatistic = this.#slots.hasSlotted('statistic');
     const hasCta = this.#slots.hasSlotted('cta');
     const isMobile = !this.#screenSize.matches.has('tabletPortrait');
+    const { on = '' } = this;
     return html`
-      <div class="${classMap({ isMobile, hasIcon, hasTitle, hasStatistic, hasCta })}">
+      <div class="${classMap({ isMobile, hasIcon, hasTitle, hasStatistic, hasCta, [on]: !!on })}">
         <span id="icon">
-          <slot name="icon" @slotchange="${this.#updateIcons}">${!this.icon ? '' : html`
-            <pfe-icon size=${this.size === 'default' ? 'md' : 'lg'} icon=${this.icon}></pfe-icon>`}
+          <slot name="icon" @slotchange="${this.#updateIcons}">${!this.icon ? '' : /* TODO: replace with rh-icon */html`
+            <pf-icon size=${this.size === 'default' ? 'md' : 'lg'}
+                     icon=${this.icon}
+                     set="${ifDefined(this.getAttribute('icon-set') ?? undefined)}"></pf-icon>`}
           </slot>
         </span>
         <span id="title"><slot name="title"></slot></span>
@@ -77,7 +80,7 @@ export class RhStat extends LitElement {
   }
 
   #updateIcons(): void {
-    this.querySelector('pfe-icon[slot="icon"]')
+    this.querySelector('pf-icon[slot="icon"]')
       ?.setAttribute?.('size', this.size === 'default' ? 'md' : 'lg');
   }