+++ elements/rh-tooltip/test/rh-tooltip.spec.ts
@@ -1,7 +1,6 @@
 import { expect, html, fixture } from '@open-wc/testing';
-import { ifDefined } from 'lit/directives/if-defined.js';
 import { RhTooltip } from '../rh-tooltip.js';
-import { setViewport } from '@web/test-runner-commands';
+import { setViewport, sendMouse } from '@web/test-runner-commands';
 
 const basicElement = html`
 <rh-tooltip position="left">
@@ -17,25 +16,10 @@ const basicElement = html`
 </rh-tooltip>
 `;
 
-const createElement = async (position: RhTooltip['position'], offset?: string) => {
-  const container = await fixture(html`
-    <div style="padding:500px;">
-      <rh-tooltip position="${position}" offset="${ifDefined(offset)}">
-        <svg xmlns="http://www.w3.org/2000/svg" height="24" width="24" viewBox="0 0 512 512">
-          <path
-            d="M256 8C119.043 8 8 119.083 8 256c0 136.997 111.043 248 248 248s248-111.003 248-248C504 119.083 392.957 8 256 8zm0 448c-110.532 0-200-89.431-200-200 0-110.495 89.472-200 200-200 110.491 0 200 89.471 200 200 0 110.53-89.431 200-200 200zm107.244-255.2c0 67.052-72.421 68.084-72.421 92.863V300c0 6.627-5.373 12-12 12h-45.647c-6.627 0-12-5.373-12-12v-8.659c0-35.745 27.1-50.034 47.579-61.516 17.561-9.845 28.324-16.541 28.324-29.579 0-17.246-21.999-28.693-39.784-28.693-23.189 0-33.894 10.977-48.942 29.969-4.057 5.12-11.46 6.071-16.666 2.124l-27.824-21.098c-5.107-3.872-6.251-11.066-2.644-16.363C184.846 131.491 214.94 112 261.794 112c49.071 0 101.45 38.304 101.45 88.8zM298 368c0 23.159-18.841 42-42 42s-42-18.841-42-42 18.841-42 42-42 42 18.841 42 42z">
-          </path>
-        </svg>
-        <div slot="content">
-          Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna
-          aliqua. Mi eget mauris pharetra et ultrices.
-        </div>
-      </rh-tooltip>
-    </div>`);
-
-  const element = container.querySelector('rh-tooltip');
+const sendMouseToTooltip = async (element: RhTooltip) => {
+  const { x, y } = element!.getBoundingClientRect();
+  await sendMouse({ position: [x, y], type: 'move' });
   await element.updateComplete;
-  return element;
 };
 
 describe('<rh-tooltip>', function() {
@@ -60,38 +44,6 @@ describe('<rh-tooltip>', function() {
     });
   });
 
-  describe('position', function() {
-    for (const position of ['left', 'top', 'right', 'bottom'] as const) {
-      describe(position, function() {
-        let placement: string;
-        beforeEach(async function() {
-          const element = await createElement(position);
-          placement = element.shadowRoot
-            ?.querySelector('[data-popper-placement]')
-            ?.getAttribute('data-popper-placement');
-        });
-        it('reflects', function() {
-          expect(placement).to.equal(position);
-        });
-      });
-    }
-  });
-
-  describe('with bottom position and offset', function() {
-    let element: RhTooltip;
-    beforeEach(async function() {
-      element = await createElement('bottom', '100, 100');
-      element.click();
-      await element.updateComplete;
-    });
-    it('should calculate an element translate', function() {
-      const style = element.shadowRoot
-        ?.querySelector('#tooltip')
-        ?.getAttribute('style');
-      expect(style).to.contain('translate');
-    });
-  });
-
   describe('with invoker and content', function() {
     let element: RhTooltip;
     beforeEach(async function() {
@@ -106,11 +58,8 @@ describe('<rh-tooltip>', function() {
     });
 
     describe('then clicking the invoker', function() {
-      beforeEach(async function() {
-        element.click();
-        await element.updateComplete;
-      });
-      it('the content should not be aria-hidden', function() {
+      it('the content should not be aria-hidden', async function() {
+        await sendMouseToTooltip(element);
         const ariaHidden = element.shadowRoot
           ?.querySelector('#tooltip')
           ?.getAttribute('aria-hidden');
@@ -119,4 +68,3 @@ describe('<rh-tooltip>', function() {
     });
   });
 });
-