+++ web-test-runner.config.js
@@ -1,11 +1,10 @@
-import { pfeTestRunnerConfig } from '@patternfly/pfe-tools/test-runner.js';
+import { pfeTestRunnerConfig } from '@patternfly/pfe-tools/test/config.js';
 import { litcssOptions } from './web-dev-server.config.js';
 
 export default pfeTestRunnerConfig({
   litcssOptions,
   tsconfig: 'tsconfig.json',
-  files: ['**/elements/**/*.spec.ts'],
+  files: ['elements/**/*.spec.ts'],
   // uncomment to get default wtr reporter
   // reporter: 'default',
 });
-